# docs/swagger.yaml
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/hello-jiangxiaoyu/QuickAuth/blob/main/LICENSE
  title: Quick Auth 接口文档
  version: "1.0"
paths:
...
  /api/quick/resources/{resourceId}/operations:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource operations
      tags:
      - resource-operation
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource operation
      tags:
      - resource-operation
  /api/quick/resources/{resourceId}/operations/{operationId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource operation
      tags:
      - resource-operation
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: get resource operation
      tags:
      - resource-operation
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource operation
      tags:
      - resource-operation
  /api/quick/resources/{resourceId}/roles:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource roles
      tags:
      - resource-role
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource role
      tags:
      - resource-role
  /api/quick/resources/{resourceId}/roles/{roleId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource role
      tags:
      - resource-role
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: get resource role
      tags:
      - resource-role
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource role
      tags:
      - resource-role
  /api/quick/resources/{resourceId}/roles/{roleId}/operations:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource role operations
      tags:
      - resource-role
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource role operation
      tags:
      - resource-role
  /api/quick/resources/{resourceId}/roles/{roleId}/operations/{operationId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource role operation
      tags:
      - resource-role
  /api/quick/resources/{resourceId}/users/{userId}/roles:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource json user roles
      tags:
      - resource-user
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource json user role
      tags:
      - resource-user
  /api/quick/resources/{resourceId}/users/{userId}/roles/{roleId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource json user role
      tags:
      - resource-user
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource json user role
      tags:
      - resource-user
...
securityDefinitions:
  Login:
    in: header
    name: token
    type: apiKey
swagger: "2.0"